requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
